refactor(users): restructure users/router for 3-layer architecture

+++ src/modules/users/router.py
@@ -1,19 +1,19 @@
-from fastapi import APIRouter, Depends, status, HTTPException
+from fastapi import APIRouter, Depends, status, HTTPException, Response
 from sqlalchemy.ext.asyncio import AsyncSession
-
 from src.db.dependencies import get_db
 from src.modules.auth.dependencies import get_current_user
 from src.modules.users.models import User
 from src.modules.users.schemas import UserResponse, UserWithTasks
-# Импортируем модуль
 from src.modules.tasks import service as task_service
 from src.modules.users import service as users_service
+from src.modules.users.service import UserNotFoundError
 
 router = APIRouter(prefix="/users", tags=["Users"])
 
 @router.get("/me", response_model=UserResponse)
 async def read_me(current_user: User = Depends(get_current_user)):
     return current_user
+
 
 @router.get("/me/tasks", response_model=UserWithTasks)
 async def read_me_with_tasks(
@@ -32,12 +32,23 @@ async def read_me_with_tasks(
 
 
 @router.delete("/{user_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_user(user_id: int, session: AsyncSession = Depends(get_db)):
-    """Удаление пользователя и всех его задач."""
-    success = await users_service.delete_user_by_id(session, user_id)
+async def delete_user(
+        user_id: int,
+        current_user: User = Depends(get_current_user),
+        session: AsyncSession = Depends(get_db)
+):
 
-    if not success:
+    if current_user.id != user_id:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="Вы можете удалить только свой собственный аккаунт"
+        )
+
+    try:
+        await users_service.delete_user_by_id(session, user_id)
+        return Response(status_code=status.HTTP_204_NO_CONTENT)
+    except UserNotFoundError as e:
         raise HTTPException(
             status_code=status.HTTP_404_NOT_FOUND,
-            detail="Пользователь не найден"
+            detail=e.message
         )
